refactor: extract server tempo/stripe config into submodules
Amp-Thread-ID: https://ampcode.com/threads/T-019d1f49-ad9f-74a8-9268-37d860491bb8
